+++ database/migrations/2019_09_16_165933_create_ssns_table.php
@@ -15,7 +15,7 @@ class CreateSsnsTable extends Migration
     {
         Schema::create('ssns', function (Blueprint $table) {
             $table->uuid('id')->primary();
-            $table->unsignedInteger('ssn_type_id');
+            $table->unsignedBigInteger('ssn_type_id');
             $table->string('ssn')->nullable();
             $table->char('number', 1)->nullable();
             $table->timestamps();
