+++ database/migrations/2019_09_27_043211_create_role_user_table.php
@@ -14,7 +14,7 @@ class CreateRoleUserTable extends Migration
     public function up()
     {
         Schema::create('role_user', function (Blueprint $table) {
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->uuid('user_id');
             $table->timestamps();
         });
